Fix some built errors and  modules
