+++ .gitignore
@@ -415,3 +415,4 @@ get_sample_data.sh
 query
 scripts/class_generators/AnnoqData_class.py
 scripts/class_generators/class_schema.json
+generated
