+++ requirements.txt
@@ -1,140 +1,15 @@
 aiohttp==3.9.3
-aiosignal==1.3.1
-annotated-types==0.6.0
-anyio==4.2.0
-async-timeout==4.0.3
-attrs==23.2.0
-certifi==2024.2.2
-click==8.1.7
-dnspython==2.5.0
-elastic-transport==8.12.0
 elasticsearch==8.12.0
-email-validator==2.1.0.post1
-exceptiongroup==1.2.0
 fastapi==0.109.2
-frozenlist==1.4.1
 graphql-core==3.2.3
-h11==0.14.0
-httpcore==1.0.2
-httptools==0.6.1
-httpx==0.26.0
-idna==3.6
-itsdangerous==2.1.2
-Jinja2==3.1.3
-libcst==1.1.0
-markdown-it-py==3.0.0
-MarkupSafe==2.1.5
-mdurl==0.1.2
-multidict==6.0.5
-mypy-extensions==1.0.0
-orjson==3.9.14
 pydantic==2.6.1
-pydantic-extra-types==2.5.0
-pydantic-settings==2.1.0
-pydantic_core==2.16.2
-Pygments==2.17.2
-python-dateutil==2.8.2
-python-dotenv==1.0.1
-python-multipart==0.0.9
-PyYAML==6.0.1
-rich==13.7.0
-six==1.16.0
-sniffio==1.3.0
-starlette==0.36.3
-strawberry-graphql==0.219.2
-typer==0.9.0
-typing-inspect==0.9.0
-typing_extensions==4.9.0
-ujson==5.9.0
-urllib3==2.2.0
+pydantic-settings
 uvicorn==0.27.1
-uvloop==0.19.0
-watchfiles==0.21.0
-websockets==12.0
-yarl==1.9.4
-aiohttp==3.9.3
-aiosignal==1.3.1
-annotated-types==0.6.0
-anyio==4.2.0
-asttokens==2.4.1
-async-timeout==4.0.3
-attrs==23.2.0
-certifi==2024.2.2
-click==8.1.7
-comm==0.2.1
-debugpy==1.8.1
-decorator==5.1.1
-dnspython==2.5.0
-elastic-transport==8.12.0
 elasticsearch==8.12.0
-email-validator==2.1.0.post1
-exceptiongroup==1.2.0
-executing==2.0.1
 fastapi==0.109.2
-frozenlist==1.4.1
 graphql-core==3.2.3
-h11==0.14.0
-httpcore==1.0.2
-httptools==0.6.1
-httpx==0.26.0
-idna==3.6
-inflection==0.5.1
-ipykernel==6.29.2
-ipython==8.21.0
-itsdangerous==2.1.2
-jedi==0.19.1
-Jinja2==3.1.3
-jsonschema==4.21.1
-jsonschema-specifications==2023.12.1
-jupyter_client==8.6.0
-jupyter_core==5.7.1
-libcst==1.1.0
-Markdown==3.5.2
-markdown-it-py==3.0.0
-MarkupSafe==2.1.5
-matplotlib-inline==0.1.6
-mdurl==0.1.2
-multidict==6.0.5
-mypy-extensions==1.0.0
-nest-asyncio==1.6.0
-orjson==3.9.14
-packaging==23.2
-parso==0.8.3
-pexpect==4.9.0
-platformdirs==4.2.0
-prompt-toolkit==3.0.43
-psutil==5.9.8
-ptyprocess==0.7.0
-pure-eval==0.2.2
 pydantic==2.6.1
-pydantic-extra-types==2.5.0
-pydantic-settings==2.1.0
-pydantic_core==2.16.2
-Pygments==2.17.2
-python-dateutil==2.8.2
-python-dotenv==1.0.1
-python-jsonschema-objects==0.5.2
-python-multipart==0.0.9
-PyYAML==6.0.1
-pyzmq==25.1.2
-referencing==0.33.0
-rich==13.7.0
-rpds-py==0.18.0
-six==1.16.0
-sniffio==1.3.0
-stack-data==0.6.3
-starlette==0.36.3
 strawberry-graphql==0.219.2
-tornado==6.4
-traitlets==5.14.1
-typer==0.9.0
-typing-inspect==0.9.0
-typing_extensions==4.9.0
-ujson==5.9.0
-urllib3==2.2.0
-uvicorn==0.27.1
-uvloop==0.19.0
-watchfiles==0.21.0
-wcwidth==0.2.13
-websockets==12.0
-yarl==1.9.4
+python-dotenv
+datamodel-code-generator
+requests
