+++ scripts/class_generators/generator.py
@@ -6,6 +6,8 @@ from elasticsearch import AsyncElasticsearch
 import os
 from dotenv import load_dotenv
 
+from utils import clean_field_name
+
 load_dotenv()
 ES_URL:str = os.getenv("ES_URL")
 ES_INDEX:str = os.getenv("ES_INDEX")
@@ -33,12 +35,7 @@ def create_class_schema_from_es_mapping():
     for key in raw_properties.keys():
         leaf = raw_properties[key]
         try:
-            name = re.sub(r'\([^)]*\)', '', key)
-            name = re.sub(r'\/[^\/]*', '', name)
-            if name[0].isdigit():
-                name = '_' + name
-            name = name.replace('-', '_')
-            name = name.replace('+', '')
+            name = clean_field_name(key)
 
             properties[name] = {"type": "model.Annotation"}
 
@@ -48,7 +45,7 @@ def create_class_schema_from_es_mapping():
            pass
 
     schema = {
-        "title": "Snps",
+        "title": "Snp",
         "type": "object", 
         "properties": {}
     }
