--- src/graphql/models/Snps.py
@@ -1,539 +1,0 @@
-# generated by datamodel-codegen:
-#   filename:  class_schema.json
-#   timestamp: 2024-03-08T00:13:43+00:00
-
-from __future__ import annotations
-
-from typing import Optional
-
-from pydantic import BaseModel
-
-from .annotation_model import Annotation
-
-
-class Snps(BaseModel):
-    x_1000Gp3_AC: Optional[Annotation] = None
-    x_1000Gp3_AF: Optional[Annotation] = None
-    x_1000Gp3_AFR_AC: Optional[Annotation] = None
-    x_1000Gp3_AFR_AF: Optional[Annotation] = None
-    x_1000Gp3_AMR_AC: Optional[Annotation] = None
-    x_1000Gp3_AMR_AF: Optional[Annotation] = None
-    x_1000Gp3_EAS_AC: Optional[Annotation] = None
-    x_1000Gp3_EAS_AF: Optional[Annotation] = None
-    x_1000Gp3_EUR_AC: Optional[Annotation] = None
-    x_1000Gp3_EUR_AF: Optional[Annotation] = None
-    x_1000Gp3_SAS_AC: Optional[Annotation] = None
-    x_1000Gp3_SAS_AF: Optional[Annotation] = None
-    ALSPAC_AC: Optional[Annotation] = None
-    ALSPAC_AF: Optional[Annotation] = None
-    ALSPAC_AN: Optional[Annotation] = None
-    ANNOVAR_ensembl_Closest_gene: Optional[Annotation] = None
-    ANNOVAR_ensembl_Effect: Optional[Annotation] = None
-    ANNOVAR_ensembl_Exon_Rank: Optional[Annotation] = None
-    ANNOVAR_ensembl_GO_biological_process_complete_list_id: Optional[Annotation] = None
-    ANNOVAR_ensembl_GO_cellular_component_complete_list_id: Optional[Annotation] = None
-    ANNOVAR_ensembl_GO_molecular_function_complete_list_id: Optional[Annotation] = None
-    ANNOVAR_ensembl_Gene_ID: Optional[Annotation] = None
-    ANNOVAR_ensembl_HGVSc: Optional[Annotation] = None
-    ANNOVAR_ensembl_HGVSp: Optional[Annotation] = None
-    ANNOVAR_ensembl_PANTHER_GO_SLIM_biological_process_list_id: Optional[Annotation] = None
-    ANNOVAR_ensembl_PANTHER_GO_SLIM_cellular_component_list_id: Optional[Annotation] = None
-    ANNOVAR_ensembl_PANTHER_GO_SLIM_molecular_function_list_id: Optional[Annotation] = None
-    ANNOVAR_ensembl_PANTHER_GO_SLIM_protein_class_list_id: Optional[Annotation] = None
-    ANNOVAR_ensembl_PANTHER_pathway_list_id: Optional[Annotation] = None
-    ANNOVAR_ensembl_REACTOME_pathway_list_id: Optional[Annotation] = None
-    ANNOVAR_ensembl_Transcript_ID: Optional[Annotation] = None
-    ANNOVAR_ensembl_summary: Optional[Annotation] = None
-    ANNOVAR_refseq_Closest_gene: Optional[Annotation] = None
-    ANNOVAR_refseq_Effect: Optional[Annotation] = None
-    ANNOVAR_refseq_Exon_Rank: Optional[Annotation] = None
-    ANNOVAR_refseq_GO_biological_process_complete_list_id: Optional[Annotation] = None
-    ANNOVAR_refseq_GO_cellular_component_complete_list_id: Optional[Annotation] = None
-    ANNOVAR_refseq_GO_molecular_function_complete_list_id: Optional[Annotation] = None
-    ANNOVAR_refseq_Gene_ID: Optional[Annotation] = None
-    ANNOVAR_refseq_HGVSc: Optional[Annotation] = None
-    ANNOVAR_refseq_HGVSp: Optional[Annotation] = None
-    ANNOVAR_refseq_PANTHER_GO_SLIM_biological_process_list_id: Optional[Annotation] = None
-    ANNOVAR_refseq_PANTHER_GO_SLIM_cellular_component_list_id: Optional[Annotation] = None
-    ANNOVAR_refseq_PANTHER_GO_SLIM_molecular_function_list_id: Optional[Annotation] = None
-    ANNOVAR_refseq_PANTHER_GO_SLIM_protein_class_list_id: Optional[Annotation] = None
-    ANNOVAR_refseq_PANTHER_pathway_list_id: Optional[Annotation] = None
-    ANNOVAR_refseq_REACTOME_pathway_list_id: Optional[Annotation] = None
-    ANNOVAR_refseq_Transcript_ID: Optional[Annotation] = None
-    ANNOVAR_refseq_summary: Optional[Annotation] = None
-    ANNOVAR_ucsc_Closest_gene: Optional[Annotation] = None
-    ANNOVAR_ucsc_Effect: Optional[Annotation] = None
-    ANNOVAR_ucsc_Exon_Rank: Optional[Annotation] = None
-    ANNOVAR_ucsc_Gene_ID: Optional[Annotation] = None
-    ANNOVAR_ucsc_HGVSc: Optional[Annotation] = None
-    ANNOVAR_ucsc_HGVSp: Optional[Annotation] = None
-    ANNOVAR_ucsc_Transcript_ID: Optional[Annotation] = None
-    ANNOVAR_ucsc_summary: Optional[Annotation] = None
-    APPRIS: Optional[Annotation] = None
-    Aloft_Confidence: Optional[Annotation] = None
-    Aloft_Fraction_transcripts_affected: Optional[Annotation] = None
-    Aloft_pred: Optional[Annotation] = None
-    Aloft_prob_Dominant: Optional[Annotation] = None
-    Aloft_prob_Recessive: Optional[Annotation] = None
-    Aloft_prob_Tolerant: Optional[Annotation] = None
-    CADD_phred: Optional[Annotation] = None
-    CADD_raw: Optional[Annotation] = None
-    CADD_raw_rankscore: Optional[Annotation] = None
-    COSMIC_CNT: Optional[Annotation] = None
-    COSMIC_ID: Optional[Annotation] = None
-    DEOGEN2_pred: Optional[Annotation] = None
-    DEOGEN2_rankscore: Optional[Annotation] = None
-    DEOGEN2_score: Optional[Annotation] = None
-    ENCODE_Dnase_cells: Optional[Annotation] = None
-    ENCODE_Dnase_score: Optional[Annotation] = None
-    ENCODE_TFBS: Optional[Annotation] = None
-    ENCODE_TFBS_cells: Optional[Annotation] = None
-    ENCODE_TFBS_score: Optional[Annotation] = None
-    ENCODE_annotated: Optional[Annotation] = None
-    ESP6500_AA_AC: Optional[Annotation] = None
-    ESP6500_AA_AF: Optional[Annotation] = None
-    ESP6500_AC: Optional[Annotation] = None
-    ESP6500_AF: Optional[Annotation] = None
-    ESP6500_EA_AC: Optional[Annotation] = None
-    ESP6500_EA_AF: Optional[Annotation] = None
-    Ensembl_Regulatory_Build_ID: Optional[Annotation] = None
-    Ensembl_Regulatory_Build_TFBS: Optional[Annotation] = None
-    Ensembl_Regulatory_Build_TFBS_matrix: Optional[Annotation] = None
-    Ensembl_Regulatory_Build_feature_type: Optional[Annotation] = None
-    Ensembl_geneid: Optional[Annotation] = None
-    Ensembl_proteinid: Optional[Annotation] = None
-    Ensembl_transcriptid: Optional[Annotation] = None
-    ExAC_AC: Optional[Annotation] = None
-    ExAC_AF: Optional[Annotation] = None
-    ExAC_AFR_AC: Optional[Annotation] = None
-    ExAC_AFR_AF: Optional[Annotation] = None
-    ExAC_AMR_AC: Optional[Annotation] = None
-    ExAC_AMR_AF: Optional[Annotation] = None
-    ExAC_Adj_AC: Optional[Annotation] = None
-    ExAC_Adj_AF: Optional[Annotation] = None
-    ExAC_EAS_AC: Optional[Annotation] = None
-    ExAC_EAS_AF: Optional[Annotation] = None
-    ExAC_FIN_AC: Optional[Annotation] = None
-    ExAC_FIN_AF: Optional[Annotation] = None
-    ExAC_NFE_AC: Optional[Annotation] = None
-    ExAC_NFE_AF: Optional[Annotation] = None
-    ExAC_SAS_AC: Optional[Annotation] = None
-    ExAC_SAS_AF: Optional[Annotation] = None
-    FANTOM5_CAGE_peak_permissive: Optional[Annotation] = None
-    FANTOM5_CAGE_peak_robust: Optional[Annotation] = None
-    FANTOM5_enhancer_differentially_expressed_tissue_cell: Optional[Annotation] = None
-    FANTOM5_enhancer_expressed_tissue_cell: Optional[Annotation] = None
-    FANTOM5_enhancer_permissive: Optional[Annotation] = None
-    FANTOM5_enhancer_robust: Optional[Annotation] = None
-    FANTOM5_enhancer_target: Optional[Annotation] = None
-    FATHMM_converted_rankscore: Optional[Annotation] = None
-    FATHMM_pred: Optional[Annotation] = None
-    FATHMM_score: Optional[Annotation] = None
-    GENCODE_basic: Optional[Annotation] = None
-    GERP_NR: Optional[Annotation] = None
-    GERP_RS: Optional[Annotation] = None
-    GERP_RS_rankscore: Optional[Annotation] = None
-    GRASP_PMID: Optional[Annotation] = None
-    GRASP_ancestry: Optional[Annotation] = None
-    GRASP_p_value: Optional[Annotation] = None
-    GRASP_phenotype: Optional[Annotation] = None
-    GRASP_platform: Optional[Annotation] = None
-    GRASP_rs: Optional[Annotation] = None
-    GTEx_V7_gene: Optional[Annotation] = None
-    GTEx_V7_tissue: Optional[Annotation] = None
-    GWAS_catalog_OR: Optional[Annotation] = None
-    GWAS_catalog_pubmedid: Optional[Annotation] = None
-    GWAS_catalog_rs: Optional[Annotation] = None
-    GWAS_catalog_trait: Optional[Annotation] = None
-    LRT_Omega: Optional[Annotation] = None
-    LRT_converted_rankscore: Optional[Annotation] = None
-    LRT_pred: Optional[Annotation] = None
-    LRT_score: Optional[Annotation] = None
-    M_CAP_pred: Optional[Annotation] = None
-    M_CAP_rankscore: Optional[Annotation] = None
-    M_CAP_score: Optional[Annotation] = None
-    MPC_rankscore: Optional[Annotation] = None
-    MPC_score: Optional[Annotation] = None
-    MVP_rankscore: Optional[Annotation] = None
-    MVP_score: Optional[Annotation] = None
-    MetaLR_pred: Optional[Annotation] = None
-    MetaLR_rankscore: Optional[Annotation] = None
-    MetaLR_score: Optional[Annotation] = None
-    MetaSVM_pred: Optional[Annotation] = None
-    MetaSVM_rankscore: Optional[Annotation] = None
-    MetaSVM_score: Optional[Annotation] = None
-    Motif_breaking: Optional[Annotation] = None
-    MutPred_AAchange: Optional[Annotation] = None
-    MutPred_Top5features: Optional[Annotation] = None
-    MutPred_protID: Optional[Annotation] = None
-    MutPred_rankscore: Optional[Annotation] = None
-    MutPred_score: Optional[Annotation] = None
-    MutationAssessor_pred: Optional[Annotation] = None
-    MutationAssessor_rankscore: Optional[Annotation] = None
-    MutationAssessor_score: Optional[Annotation] = None
-    MutationTaster_AAE: Optional[Annotation] = None
-    MutationTaster_converted_rankscore: Optional[Annotation] = None
-    MutationTaster_model: Optional[Annotation] = None
-    MutationTaster_pred: Optional[Annotation] = None
-    MutationTaster_score: Optional[Annotation] = None
-    ORegAnno_PMID: Optional[Annotation] = None
-    ORegAnno_type: Optional[Annotation] = None
-    PROVEAN_converted_rankscore: Optional[Annotation] = None
-    PROVEAN_pred: Optional[Annotation] = None
-    PROVEAN_score: Optional[Annotation] = None
-    PrimateAI_pred: Optional[Annotation] = None
-    PrimateAI_rankscore: Optional[Annotation] = None
-    PrimateAI_score: Optional[Annotation] = None
-    RegulomeDB_motif: Optional[Annotation] = None
-    RegulomeDB_score: Optional[Annotation] = None
-    Reliability_index: Optional[Annotation] = None
-    SIFT4G_converted_rankscore: Optional[Annotation] = None
-    SIFT4G_pred: Optional[Annotation] = None
-    SIFT4G_score: Optional[Annotation] = None
-    SIFT_converted_rankscore: Optional[Annotation] = None
-    SIFT_pred: Optional[Annotation] = None
-    SIFT_score: Optional[Annotation] = None
-    SnpEff_ensembl_CDS_position: Optional[Annotation] = None
-    SnpEff_ensembl_Distance_to_feature: Optional[Annotation] = None
-    SnpEff_ensembl_Effect: Optional[Annotation] = None
-    SnpEff_ensembl_Effect_impact: Optional[Annotation] = None
-    SnpEff_ensembl_Exon_or_intron_rank: Optional[Annotation] = None
-    SnpEff_ensembl_GO_biological_process_complete_list_id: Optional[Annotation] = None
-    SnpEff_ensembl_GO_cellular_component_complete_list_id: Optional[Annotation] = None
-    SnpEff_ensembl_GO_molecular_function_complete_list_id: Optional[Annotation] = None
-    SnpEff_ensembl_Gene_ID: Optional[Annotation] = None
-    SnpEff_ensembl_Gene_name: Optional[Annotation] = None
-    SnpEff_ensembl_HGVSc: Optional[Annotation] = None
-    SnpEff_ensembl_HGVSp: Optional[Annotation] = None
-    SnpEff_ensembl_LOF: Optional[Annotation] = None
-    SnpEff_ensembl_PANTHER_GO_SLIM_biological_process_list_id: Optional[Annotation] = None
-    SnpEff_ensembl_PANTHER_GO_SLIM_cellular_component_list_id: Optional[Annotation] = None
-    SnpEff_ensembl_PANTHER_GO_SLIM_molecular_function_list_id: Optional[Annotation] = None
-    SnpEff_ensembl_PANTHER_GO_SLIM_protein_class_list_id: Optional[Annotation] = None
-    SnpEff_ensembl_PANTHER_pathway_list_id: Optional[Annotation] = None
-    SnpEff_ensembl_Protein_position: Optional[Annotation] = None
-    SnpEff_ensembl_REACTOME_pathway_list_id: Optional[Annotation] = None
-    SnpEff_ensembl_Sequence_feature: Optional[Annotation] = None
-    SnpEff_ensembl_Sequence_feature_impact: Optional[Annotation] = None
-    SnpEff_ensembl_TF_ID: Optional[Annotation] = None
-    SnpEff_ensembl_TF_binding_effect: Optional[Annotation] = None
-    SnpEff_ensembl_TF_name: Optional[Annotation] = None
-    SnpEff_ensembl_Transcript_ID: Optional[Annotation] = None
-    SnpEff_ensembl_Transcript_biotype: Optional[Annotation] = None
-    SnpEff_ensembl_Warnings: Optional[Annotation] = None
-    SnpEff_ensembl_cDNA_position: Optional[Annotation] = None
-    SnpEff_ensembl_summary: Optional[Annotation] = None
-    SnpEff_refseq_CDS_position: Optional[Annotation] = None
-    SnpEff_refseq_Distance_to_feature: Optional[Annotation] = None
-    SnpEff_refseq_Effect: Optional[Annotation] = None
-    SnpEff_refseq_Effect_impact: Optional[Annotation] = None
-    SnpEff_refseq_Exon_or_intron_rank: Optional[Annotation] = None
-    SnpEff_refseq_GO_biological_process_complete_list_id: Optional[Annotation] = None
-    SnpEff_refseq_GO_cellular_component_complete_list_id: Optional[Annotation] = None
-    SnpEff_refseq_GO_molecular_function_complete_list_id: Optional[Annotation] = None
-    SnpEff_refseq_Gene_ID: Optional[Annotation] = None
-    SnpEff_refseq_Gene_name: Optional[Annotation] = None
-    SnpEff_refseq_HGVSc: Optional[Annotation] = None
-    SnpEff_refseq_HGVSp: Optional[Annotation] = None
-    SnpEff_refseq_LOF: Optional[Annotation] = None
-    SnpEff_refseq_PANTHER_GO_SLIM_biological_process_list_id: Optional[Annotation] = None
-    SnpEff_refseq_PANTHER_GO_SLIM_cellular_component_list_id: Optional[Annotation] = None
-    SnpEff_refseq_PANTHER_GO_SLIM_molecular_function_list_id: Optional[Annotation] = None
-    SnpEff_refseq_PANTHER_GO_SLIM_protein_class_list_id: Optional[Annotation] = None
-    SnpEff_refseq_PANTHER_pathway_list_id: Optional[Annotation] = None
-    SnpEff_refseq_Protein_position: Optional[Annotation] = None
-    SnpEff_refseq_REACTOME_pathway_list_id: Optional[Annotation] = None
-    SnpEff_refseq_Sequence_feature: Optional[Annotation] = None
-    SnpEff_refseq_Sequence_feature_impact: Optional[Annotation] = None
-    SnpEff_refseq_Transcript_ID: Optional[Annotation] = None
-    SnpEff_refseq_Transcript_biotype: Optional[Annotation] = None
-    SnpEff_refseq_Warnings: Optional[Annotation] = None
-    SnpEff_refseq_cDNA_position: Optional[Annotation] = None
-    SnpEff_refseq_summary: Optional[Annotation] = None
-    TSL: Optional[Annotation] = None
-    TWINSUK_AC: Optional[Annotation] = None
-    TWINSUK_AF: Optional[Annotation] = None
-    TWINSUK_AN: Optional[Annotation] = None
-    TargetScan_context_score_percentile: Optional[Annotation] = None
-    UK10K_AC: Optional[Annotation] = None
-    UK10K_AF: Optional[Annotation] = None
-    UK10K_AN: Optional[Annotation] = None
-    UTR3_miRNA_target: Optional[Annotation] = None
-    Uniprot_acc: Optional[Annotation] = None
-    Uniprot_entry: Optional[Annotation] = None
-    VEP_canonical: Optional[Annotation] = None
-    VEP_ensembl_Amino_Acid_Change: Optional[Annotation] = None
-    VEP_ensembl_CANONICAL: Optional[Annotation] = None
-    VEP_ensembl_CCDS: Optional[Annotation] = None
-    VEP_ensembl_CDS_position: Optional[Annotation] = None
-    VEP_ensembl_Codon_Change_or_Distance: Optional[Annotation] = None
-    VEP_ensembl_Consequence: Optional[Annotation] = None
-    VEP_ensembl_Exon_or_Intron_Rank: Optional[Annotation] = None
-    VEP_ensembl_GO_biological_process_complete_list_id: Optional[Annotation] = None
-    VEP_ensembl_GO_cellular_component_complete_list_id: Optional[Annotation] = None
-    VEP_ensembl_GO_molecular_function_complete_list_id: Optional[Annotation] = None
-    VEP_ensembl_Gene_ID: Optional[Annotation] = None
-    VEP_ensembl_Gene_Name: Optional[Annotation] = None
-    VEP_ensembl_HGVSc: Optional[Annotation] = None
-    VEP_ensembl_HGVSp: Optional[Annotation] = None
-    VEP_ensembl_LoF: Optional[Annotation] = None
-    VEP_ensembl_LoF_filter: Optional[Annotation] = None
-    VEP_ensembl_LoF_flags: Optional[Annotation] = None
-    VEP_ensembl_LoF_info: Optional[Annotation] = None
-    VEP_ensembl_PANTHER_GO_SLIM_biological_process_list_id: Optional[Annotation] = None
-    VEP_ensembl_PANTHER_GO_SLIM_cellular_component_list_id: Optional[Annotation] = None
-    VEP_ensembl_PANTHER_GO_SLIM_molecular_function_list_id: Optional[Annotation] = None
-    VEP_ensembl_PANTHER_GO_SLIM_protein_class_list_id: Optional[Annotation] = None
-    VEP_ensembl_PANTHER_pathway_list_id: Optional[Annotation] = None
-    VEP_ensembl_Protein_ID: Optional[Annotation] = None
-    VEP_ensembl_Protein_position: Optional[Annotation] = None
-    VEP_ensembl_REACTOME_pathway_list_id: Optional[Annotation] = None
-    VEP_ensembl_STRAND: Optional[Annotation] = None
-    VEP_ensembl_SWISSPROT: Optional[Annotation] = None
-    VEP_ensembl_Transcript_ID: Optional[Annotation] = None
-    VEP_ensembl_cDNA_position: Optional[Annotation] = None
-    VEP_ensembl_summary: Optional[Annotation] = None
-    VEP_refseq_Amino_Acid_Change: Optional[Annotation] = None
-    VEP_refseq_CANONICAL: Optional[Annotation] = None
-    VEP_refseq_CDS_position: Optional[Annotation] = None
-    VEP_refseq_Codon_Change_or_Distance: Optional[Annotation] = None
-    VEP_refseq_Consequence: Optional[Annotation] = None
-    VEP_refseq_Exon_or_Intron_Rank: Optional[Annotation] = None
-    VEP_refseq_GO_biological_process_complete_list_id: Optional[Annotation] = None
-    VEP_refseq_GO_cellular_component_complete_list_id: Optional[Annotation] = None
-    VEP_refseq_GO_molecular_function_complete_list_id: Optional[Annotation] = None
-    VEP_refseq_Gene_ID: Optional[Annotation] = None
-    VEP_refseq_Gene_Name: Optional[Annotation] = None
-    VEP_refseq_HGVSc: Optional[Annotation] = None
-    VEP_refseq_HGVSp: Optional[Annotation] = None
-    VEP_refseq_LoF: Optional[Annotation] = None
-    VEP_refseq_LoF_filter: Optional[Annotation] = None
-    VEP_refseq_LoF_flags: Optional[Annotation] = None
-    VEP_refseq_LoF_info: Optional[Annotation] = None
-    VEP_refseq_PANTHER_GO_SLIM_biological_process_list_id: Optional[Annotation] = None
-    VEP_refseq_PANTHER_GO_SLIM_cellular_component_list_id: Optional[Annotation] = None
-    VEP_refseq_PANTHER_GO_SLIM_molecular_function_list_id: Optional[Annotation] = None
-    VEP_refseq_PANTHER_GO_SLIM_protein_class_list_id: Optional[Annotation] = None
-    VEP_refseq_PANTHER_pathway_list_id: Optional[Annotation] = None
-    VEP_refseq_Protein_ID: Optional[Annotation] = None
-    VEP_refseq_Protein_position: Optional[Annotation] = None
-    VEP_refseq_REACTOME_pathway_list_id: Optional[Annotation] = None
-    VEP_refseq_STRAND: Optional[Annotation] = None
-    VEP_refseq_Transcript_ID: Optional[Annotation] = None
-    VEP_refseq_cDNA_position: Optional[Annotation] = None
-    VEP_refseq_summary: Optional[Annotation] = None
-    aaalt: Optional[Annotation] = None
-    aapos: Optional[Annotation] = None
-    aaref: Optional[Annotation] = None
-    alt: Optional[Annotation] = None
-    cds_strand: Optional[Annotation] = None
-    chr: Optional[Annotation] = None
-    clinvar_MedGen_id: Optional[Annotation] = None
-    clinvar_OMIM_id: Optional[Annotation] = None
-    clinvar_Orphanet_id: Optional[Annotation] = None
-    clinvar_clnsig: Optional[Annotation] = None
-    clinvar_hgvs: Optional[Annotation] = None
-    clinvar_id: Optional[Annotation] = None
-    clinvar_review: Optional[Annotation] = None
-    clinvar_trait: Optional[Annotation] = None
-    clinvar_var_source: Optional[Annotation] = None
-    codon_degeneracy: Optional[Annotation] = None
-    codonpos: Optional[Annotation] = None
-    enhancer_linked_GO_biological_process_complete_list_id: Optional[Annotation] = None
-    enhancer_linked_GO_cellular_component_complete_list_id: Optional[Annotation] = None
-    enhancer_linked_GO_molecular_function_complete_list_id: Optional[Annotation] = None
-    enhancer_linked_PANTHER_GO_SLIM_biological_process_list_id: Optional[Annotation] = None
-    enhancer_linked_PANTHER_GO_SLIM_cellular_component_list_id: Optional[Annotation] = None
-    enhancer_linked_PANTHER_GO_SLIM_molecular_function_list_id: Optional[Annotation] = None
-    enhancer_linked_REACTOME_pathway_list_id: Optional[Annotation] = None
-    enhancer_linked_enhancer: Optional[Annotation] = None
-    enhancer_linked_genes: Optional[Annotation] = None
-    fathmm_MKL_coding_group: Optional[Annotation] = None
-    fathmm_MKL_coding_pred: Optional[Annotation] = None
-    fathmm_MKL_coding_rankscore: Optional[Annotation] = None
-    fathmm_MKL_coding_score: Optional[Annotation] = None
-    fathmm_MKL_non_coding_group: Optional[Annotation] = None
-    fathmm_MKL_non_coding_pred: Optional[Annotation] = None
-    fathmm_MKL_non_coding_rankscore: Optional[Annotation] = None
-    fathmm_MKL_non_coding_score: Optional[Annotation] = None
-    fathmm_XF_coding_or_noncoding: Optional[Annotation] = None
-    fathmm_XF_pred: Optional[Annotation] = None
-    fathmm_XF_rankscore: Optional[Annotation] = None
-    fathmm_XF_score: Optional[Annotation] = None
-    flanking_0_GO_biological_process_complete_list_id: Optional[Annotation] = None
-    flanking_0_GO_cellular_component_complete_list_id: Optional[Annotation] = None
-    flanking_0_GO_molecular_function_complete_list_id: Optional[Annotation] = None
-    flanking_0_PANTHER_GO_SLIM_biological_process_list_id: Optional[Annotation] = None
-    flanking_0_PANTHER_GO_SLIM_cellular_component_list_id: Optional[Annotation] = None
-    flanking_0_PANTHER_GO_SLIM_molecular_function_list_id: Optional[Annotation] = None
-    flanking_0_PANTHER_pathway_list_id: Optional[Annotation] = None
-    flanking_0_REACTOME_pathway_list_id: Optional[Annotation] = None
-    flanking_10000_GO_biological_process_complete_list_id: Optional[Annotation] = None
-    flanking_10000_GO_cellular_component_complete_list_id: Optional[Annotation] = None
-    flanking_10000_GO_molecular_function_complete_list_id: Optional[Annotation] = None
-    flanking_10000_PANTHER_GO_SLIM_biological_process_list_id: Optional[Annotation] = None
-    flanking_10000_PANTHER_GO_SLIM_cellular_component_list_id: Optional[Annotation] = None
-    flanking_10000_PANTHER_GO_SLIM_molecular_function_list_id: Optional[Annotation] = None
-    flanking_10000_PANTHER_pathway_list_id: Optional[Annotation] = None
-    flanking_10000_REACTOME_pathway_list_id: Optional[Annotation] = None
-    flanking_20000_GO_biological_process_complete_list_id: Optional[Annotation] = None
-    flanking_20000_GO_cellular_component_complete_list_id: Optional[Annotation] = None
-    flanking_20000_GO_molecular_function_complete_list_id: Optional[Annotation] = None
-    flanking_20000_PANTHER_GO_SLIM_biological_process_list_id: Optional[Annotation] = None
-    flanking_20000_PANTHER_GO_SLIM_cellular_component_list_id: Optional[Annotation] = None
-    flanking_20000_PANTHER_GO_SLIM_molecular_function_list_id: Optional[Annotation] = None
-    flanking_20000_PANTHER_GO_SLIM_protein_class_list_id: Optional[Annotation] = None
-    flanking_20000_PANTHER_pathway_list_id: Optional[Annotation] = None
-    flanking_20000_REACTOME_pathway_list_id: Optional[Annotation] = None
-    funseq_noncoding_score: Optional[Annotation] = None
-    genename: Optional[Annotation] = None
-    gnomAD_exomes_AC: Optional[Annotation] = None
-    gnomAD_exomes_AF: Optional[Annotation] = None
-    gnomAD_exomes_AFR_AC: Optional[Annotation] = None
-    gnomAD_exomes_AFR_AF: Optional[Annotation] = None
-    gnomAD_exomes_AFR_AN: Optional[Annotation] = None
-    gnomAD_exomes_AFR_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_AMR_AC: Optional[Annotation] = None
-    gnomAD_exomes_AMR_AF: Optional[Annotation] = None
-    gnomAD_exomes_AMR_AN: Optional[Annotation] = None
-    gnomAD_exomes_AMR_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_AN: Optional[Annotation] = None
-    gnomAD_exomes_ASJ_AC: Optional[Annotation] = None
-    gnomAD_exomes_ASJ_AF: Optional[Annotation] = None
-    gnomAD_exomes_ASJ_AN: Optional[Annotation] = None
-    gnomAD_exomes_ASJ_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_EAS_AC: Optional[Annotation] = None
-    gnomAD_exomes_EAS_AF: Optional[Annotation] = None
-    gnomAD_exomes_EAS_AN: Optional[Annotation] = None
-    gnomAD_exomes_EAS_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_FIN_AC: Optional[Annotation] = None
-    gnomAD_exomes_FIN_AF: Optional[Annotation] = None
-    gnomAD_exomes_FIN_AN: Optional[Annotation] = None
-    gnomAD_exomes_FIN_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_NFE_AC: Optional[Annotation] = None
-    gnomAD_exomes_NFE_AF: Optional[Annotation] = None
-    gnomAD_exomes_NFE_AN: Optional[Annotation] = None
-    gnomAD_exomes_NFE_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_POPMAX_AC: Optional[Annotation] = None
-    gnomAD_exomes_POPMAX_AF: Optional[Annotation] = None
-    gnomAD_exomes_POPMAX_AN: Optional[Annotation] = None
-    gnomAD_exomes_POPMAX_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_SAS_AC: Optional[Annotation] = None
-    gnomAD_exomes_SAS_AF: Optional[Annotation] = None
-    gnomAD_exomes_SAS_AN: Optional[Annotation] = None
-    gnomAD_exomes_SAS_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_controls_AC: Optional[Annotation] = None
-    gnomAD_exomes_controls_AF: Optional[Annotation] = None
-    gnomAD_exomes_controls_AFR_AC: Optional[Annotation] = None
-    gnomAD_exomes_controls_AFR_AF: Optional[Annotation] = None
-    gnomAD_exomes_controls_AFR_AN: Optional[Annotation] = None
-    gnomAD_exomes_controls_AFR_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_controls_AMR_AC: Optional[Annotation] = None
-    gnomAD_exomes_controls_AMR_AF: Optional[Annotation] = None
-    gnomAD_exomes_controls_AMR_AN: Optional[Annotation] = None
-    gnomAD_exomes_controls_AMR_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_controls_AN: Optional[Annotation] = None
-    gnomAD_exomes_controls_ASJ_AC: Optional[Annotation] = None
-    gnomAD_exomes_controls_ASJ_AF: Optional[Annotation] = None
-    gnomAD_exomes_controls_ASJ_AN: Optional[Annotation] = None
-    gnomAD_exomes_controls_ASJ_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_controls_EAS_AC: Optional[Annotation] = None
-    gnomAD_exomes_controls_EAS_AF: Optional[Annotation] = None
-    gnomAD_exomes_controls_EAS_AN: Optional[Annotation] = None
-    gnomAD_exomes_controls_EAS_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_controls_FIN_AC: Optional[Annotation] = None
-    gnomAD_exomes_controls_FIN_AF: Optional[Annotation] = None
-    gnomAD_exomes_controls_FIN_AN: Optional[Annotation] = None
-    gnomAD_exomes_controls_FIN_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_controls_NFE_AC: Optional[Annotation] = None
-    gnomAD_exomes_controls_NFE_AF: Optional[Annotation] = None
-    gnomAD_exomes_controls_NFE_AN: Optional[Annotation] = None
-    gnomAD_exomes_controls_NFE_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_controls_POPMAX_AC: Optional[Annotation] = None
-    gnomAD_exomes_controls_POPMAX_AF: Optional[Annotation] = None
-    gnomAD_exomes_controls_POPMAX_AN: Optional[Annotation] = None
-    gnomAD_exomes_controls_POPMAX_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_controls_SAS_AC: Optional[Annotation] = None
-    gnomAD_exomes_controls_SAS_AF: Optional[Annotation] = None
-    gnomAD_exomes_controls_SAS_AN: Optional[Annotation] = None
-    gnomAD_exomes_controls_SAS_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_controls_nhomalt: Optional[Annotation] = None
-    gnomAD_exomes_flag: Optional[Annotation] = None
-    gnomAD_exomes_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_AC: Optional[Annotation] = None
-    gnomAD_genomes_AF: Optional[Annotation] = None
-    gnomAD_genomes_AFR_AC: Optional[Annotation] = None
-    gnomAD_genomes_AFR_AF: Optional[Annotation] = None
-    gnomAD_genomes_AFR_AN: Optional[Annotation] = None
-    gnomAD_genomes_AFR_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_AMR_AC: Optional[Annotation] = None
-    gnomAD_genomes_AMR_AF: Optional[Annotation] = None
-    gnomAD_genomes_AMR_AN: Optional[Annotation] = None
-    gnomAD_genomes_AMR_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_AN: Optional[Annotation] = None
-    gnomAD_genomes_ASJ_AC: Optional[Annotation] = None
-    gnomAD_genomes_ASJ_AF: Optional[Annotation] = None
-    gnomAD_genomes_ASJ_AN: Optional[Annotation] = None
-    gnomAD_genomes_ASJ_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_EAS_AC: Optional[Annotation] = None
-    gnomAD_genomes_EAS_AF: Optional[Annotation] = None
-    gnomAD_genomes_EAS_AN: Optional[Annotation] = None
-    gnomAD_genomes_EAS_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_FIN_AC: Optional[Annotation] = None
-    gnomAD_genomes_FIN_AF: Optional[Annotation] = None
-    gnomAD_genomes_FIN_AN: Optional[Annotation] = None
-    gnomAD_genomes_FIN_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_NFE_AC: Optional[Annotation] = None
-    gnomAD_genomes_NFE_AF: Optional[Annotation] = None
-    gnomAD_genomes_NFE_AN: Optional[Annotation] = None
-    gnomAD_genomes_NFE_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_POPMAX_AC: Optional[Annotation] = None
-    gnomAD_genomes_POPMAX_AF: Optional[Annotation] = None
-    gnomAD_genomes_POPMAX_AN: Optional[Annotation] = None
-    gnomAD_genomes_POPMAX_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_controls_AC: Optional[Annotation] = None
-    gnomAD_genomes_controls_AF: Optional[Annotation] = None
-    gnomAD_genomes_controls_AFR_AC: Optional[Annotation] = None
-    gnomAD_genomes_controls_AFR_AF: Optional[Annotation] = None
-    gnomAD_genomes_controls_AFR_AN: Optional[Annotation] = None
-    gnomAD_genomes_controls_AFR_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_controls_AMR_AC: Optional[Annotation] = None
-    gnomAD_genomes_controls_AMR_AF: Optional[Annotation] = None
-    gnomAD_genomes_controls_AMR_AN: Optional[Annotation] = None
-    gnomAD_genomes_controls_AMR_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_controls_AN: Optional[Annotation] = None
-    gnomAD_genomes_controls_ASJ_AC: Optional[Annotation] = None
-    gnomAD_genomes_controls_ASJ_AF: Optional[Annotation] = None
-    gnomAD_genomes_controls_ASJ_AN: Optional[Annotation] = None
-    gnomAD_genomes_controls_ASJ_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_controls_EAS_AC: Optional[Annotation] = None
-    gnomAD_genomes_controls_EAS_AF: Optional[Annotation] = None
-    gnomAD_genomes_controls_EAS_AN: Optional[Annotation] = None
-    gnomAD_genomes_controls_EAS_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_controls_FIN_AC: Optional[Annotation] = None
-    gnomAD_genomes_controls_FIN_AF: Optional[Annotation] = None
-    gnomAD_genomes_controls_FIN_AN: Optional[Annotation] = None
-    gnomAD_genomes_controls_FIN_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_controls_NFE_AC: Optional[Annotation] = None
-    gnomAD_genomes_controls_NFE_AF: Optional[Annotation] = None
-    gnomAD_genomes_controls_NFE_AN: Optional[Annotation] = None
-    gnomAD_genomes_controls_NFE_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_controls_POPMAX_AC: Optional[Annotation] = None
-    gnomAD_genomes_controls_POPMAX_AF: Optional[Annotation] = None
-    gnomAD_genomes_controls_POPMAX_AN: Optional[Annotation] = None
-    gnomAD_genomes_controls_POPMAX_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_controls_nhomalt: Optional[Annotation] = None
-    gnomAD_genomes_flag: Optional[Annotation] = None
-    gnomAD_genomes_nhomalt: Optional[Annotation] = None
-    network_hub: Optional[Annotation] = None
-    pos: Optional[Annotation] = None
-    ref: Optional[Annotation] = None
-    refcodon: Optional[Annotation] = None
-    rs_dbSNP151: Optional[Annotation] = None
-    sensitive: Optional[Annotation] = None
-    sno_miRNA_name: Optional[Annotation] = None
-    sno_miRNA_type: Optional[Annotation] = None
-    splicing_consensus_ada_score: Optional[Annotation] = None
-    splicing_consensus_rf_score: Optional[Annotation] = None
-    target_gene: Optional[Annotation] = None
-    ultra_sensitive: Optional[Annotation] = None
