+++ src/graphql/resolvers/helper_resolver.py
@@ -1,24 +1,15 @@
-from ..gene_pos import get_pos_from_gene_id, map_gene, chromosomal_location_dic
-from ..models.snp_model import SnpsType
-from ..models.annotation_model import AggregationItem, Bucket, DocCount, Annotation
+from src.graphql.gene_pos import get_pos_from_gene_id, map_gene, chromosomal_location_dic
+from src.graphql.models.snp_model import SnpsType
+from src.graphql.models.annotation_model import AggregationItem, Bucket, DocCount, Annotation
 import re
 
-
-def to_graphql_name(name):
-    if name[0].isdigit():
-        return f"x_{name}"
-    name = re.sub(r'\([^)]*\)', '', name)
-    name = re.sub(r'\/[^\/]*', '', name)
-    name = name.replace('-', '_')
-    name = name.replace('+', '')
-    return name
-
+from utils import clean_field_name
 
 def convert_hits(hits, aggregations):
     compliant_results = []
     for hit in hits:
         source = hit['_source']
-        compliant_source = {to_graphql_name(key): value for key, value in source.items()}
+        compliant_source = {clean_field_name(key): value for key, value in source.items()}
 
         data = {}
