+++ src/graphql/resolvers/snp_resolver.py
@@ -1,7 +1,7 @@
-from ...config.es import es
-from ...config.settings import settings
-from ..models.annotation_model import PageArgs
-from .helper_resolver import IDs_query, annotation_query, chromosome_query, convert_hits, gene_query, get_aggregation_query, rsID_query, rsIDs_query
+from src.config.es import es
+from src.config.settings import settings
+from src.graphql.models.annotation_model import PageArgs
+from src.graphql.resolvers.helper_resolver import IDs_query, annotation_query, chromosome_query, convert_hits, gene_query, get_aggregation_query, rsID_query, rsIDs_query
 
 
 # Query for getting all annotations, no filter, size 20
